requirements .txt
streamlit
Pillow
google-generativeai
Datetime



